Consolidate plotting interfaces and functions.
