+++ xcp_d/interfaces/__init__.py
@@ -9,13 +9,13 @@ from xcp_d.interfaces import (
     filtering,
     layout_builder,
     nilearn,
+    plotting,
     prepostcleaning,
-    qc_plot,
     regression,
     report,
     report_core,
     resting_state,
-    surfplotting,
+    plotting,
     workbench,
 )
 
@@ -28,11 +28,11 @@ __all__ = [
     "layout_builder",
     "nilearn",
     "prepostcleaning",
-    "qc_plot",
+    "plotting",
     "regression",
     "report_core",
     "report",
     "resting_state",
-    "surfplotting",
+    "plotting",
     "workbench",
 ]
