+++ xcp_d/workflow/bold.py
@@ -29,7 +29,7 @@ from xcp_d.utils.confounds import (
 )
 from xcp_d.utils.doc import fill_doc
 from xcp_d.utils.filemanip import check_binary_mask
-from xcp_d.utils.plot import plot_design_matrix
+from xcp_d.utils.plotting import plot_design_matrix
 from xcp_d.workflow.connectivity import init_nifti_functional_connectivity_wf
 from xcp_d.workflow.execsummary import init_execsummary_wf
 from xcp_d.workflow.outputs import init_writederivatives_wf
